+++ src/Sentry.OpenTelemetry/IOpenTelemetryEnricher.cs
@@ -2,5 +2,5 @@ namespace Sentry.OpenTelemetry;
 
 internal interface IOpenTelemetryEnricher
 {
-    void Enrich(ISpan span, Activity activity, IHub hub, SentryOptions? options);
+    void Enrich(ISpanTracer span, Activity activity, IHub hub, SentryOptions? options);
 }